Memoria: Don't redirect if no seed
Instead, just provide a default seed.

Yes, now refreshing the page rolls the dice, but it does not matter when the game is not started yet. Unless someone wants to share exactly the same game seed with a friend, but that should probably be properly implemented as a feature and suggested in the UI.